requirements.txt
web3==6.11.1
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.1
cryptography==41.0.7
requests==2.31.0
python-json-logger==2.0.7 